requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
yt-dlp==2025.6.9
python-dotenv==1.0.1